[refactor] Change DIR
related to #110
