+++ search/src/main/java/com/bjcareer/search/candidate/cache/CacheTrieService.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 import org.bson.Document;
 
 import com.bjcareer.search.candidate.Trie;
-import com.bjcareer.search.out.persistence.repository.cache.CacheRepository;
-import com.bjcareer.search.out.persistence.repository.noSQL.DocumentRepository;
+import com.bjcareer.search.out.persistence.cache.CacheRepository;
+import com.bjcareer.search.out.persistence.noSQL.DocumentRepository;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
