+++ search/src/main/java/com/bjcareer/search/schedule/ScheduleOhlcService.java
@@ -20,7 +20,7 @@ import com.bjcareer.search.config.AppConfig;
 import com.bjcareer.search.domain.entity.Market;
 import com.bjcareer.search.domain.entity.Stock;
 import com.bjcareer.search.domain.entity.StockChart;
-import com.bjcareer.search.out.persistence.repository.stock.StockRepositoryAdapter;
+import com.bjcareer.search.out.persistence.stock.StockRepositoryAdapter;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
