+++ search/src/main/java/com/bjcareer/search/schedule/ScheduleTrieService.java
@@ -3,12 +3,11 @@ package com.bjcareer.search.schedule;
 import java.util.List;
 
 import org.bson.Document;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import com.bjcareer.search.candidate.Trie;
 import com.bjcareer.search.candidate.noSQL.DocumentQueryKeywords;
-import com.bjcareer.search.out.persistence.repository.noSQL.DocumentRepository;
+import com.bjcareer.search.out.persistence.noSQL.DocumentRepository;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
