+++ search/src/test/java/com/bjcareer/search/candidate/cache/CacheTrieServiceTest.java
@@ -14,8 +14,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-import com.bjcareer.search.out.persistence.repository.cache.CacheRepository;
-import com.bjcareer.search.out.persistence.repository.noSQL.DocumentRepository;
+import com.bjcareer.search.out.persistence.cache.CacheRepository;
+import com.bjcareer.search.out.persistence.noSQL.DocumentRepository;
 
 class CacheTrieServiceTest {
